Fix logic error in MultiAppNearestNodeTransfer.
It was not originally possible for MultiAppNearestNodeTransfer::bboxMinDistance()
to return anything other than 0, and this was causing the _first_
candidate BBox tested for "nearness" to always be accepted. It was
therefore possible that we were getting values Transferred which were
not technically from the nearest SubApp. It looks like the faulty
logic was originally added in a9f6751bb46, so it has been this way for
quite a while...

I have also improved the error checking of this Transfer. In
principle, it should not be possible for _no_ candidate BoundingBoxes
to be found for a given Point, but we were not actually
checking/asserting this fact previously, and I ran into issues while
trying to debug it that could have been prevented with some
simple/cheap extra checks.  These extra checks have now been added.

There is also an issue with the entire concept of the
bboxMin/MaxDistance() functions, since instead of measuring the "true"
distance from a given point to a BoundingBox, they are measuring the
distance from a given point to each of the BBox "corners". Since a
point can be much "closer" to a BBox's side than it is to any of its
corners, using the distance to a corner can give a very skewed idea of
which BBox might be the best or "closest" candidate for a nearest
node. I am planning to address this issue in an upcoming PR, for now I
want to make sure that all the MOOSE apps are happy with the current
logic fix and address/fix any diffs that arise.

* I also improved the documentation of and slightly modified the
  two_way_many_apps_master.i test so that there is no ambiguity about
  which SubApp is closest to a given Elem centroid. This makes it easier
  to visually determine whether the output of this particular test is
  what we expect it to be, and means that the results are not subject to
  changes in SubApp ordering, etc.
* Add design/issues/requirement files to all NearestNodeTransfer tests.
* Fix typos in some test documentation.

Refs #2126