fix segfault in certain dynamic loading scenarios
When dynamic loading/registration is used, MOOSE calls dlclose in the
MooseApp constructor.  So all objects created from dlopen'd libs need to
be destroyed before that call.  The new warehouse now holds shared
pointers to objects that would previously have been deallocated earlier.
To fix this, make the new master warehouse a unique pointer that we
explicitly deallocate in the MooseApp destructor *before* the dlclose
call.

ref #11994