Correctly compute volumetric_strain using logarithmic strains closes #11743
In both solid_mechanics and tensor_mechanics, the volumetric_strain
quantity computed from the rank 2 tensor was computed assuming that the
strain was an engineering strain, when in reality it is a logarithmic
strain in both systems. This changes the calculation to use logarithmic
strains, and tests to ensure that the computed quantity is the same in
equivalent models in both systems.

Note that one modules test diffed substantially with this change
(internal_volume.rz_displaced), and its new volumetric_strain is now
consistent with the computed internal volume.