fix Attribute comparison
Previously the same function was used for doing query matches on
attributes and for comparing attribute equality (e.g. as used by stl
containers like map, set, unordered map, etc.).  This was wrong.  Hash
collisions resulted in equality comparisons that returned true for
non-identical attribute instances because the query-comparison (isMatch)
was used.  This resulted in the wrong query id being returned for
certain queries (one that was query-equivalent but not identical).

fixes #12368, ref #11994