Handle degenerate source case
If a source processor contains no part of the mesh, we don't want the bounding box of the domain
to get in the way when deciding where to find points during our nearest entity search. This commit
adds a method which filters the domain to only include the source boundary nodes.