LLVM/MPICH is hard
Alert the user they are treading in difficult waters when choosing to build the LLVM/MPICH stack. Because it's hard. Really. I barely understand it myself.
Add a sort of page break between adjacent paragraph siblings within our alerts.
Add instructions on skipping MPICH for GCC if its not needed.