Redoing how HeatStructure is coupled to a Pipe
Now, we find out which elements from a Pipe correspond to HeatStructure
elements. Then, we transfer the surface temperature from the heat structure
into a aux variable in the Pipe. We sweep across the pipe and compute heat
flux which is then cached in a user object.  Then, we use a BC object and
a Kernel object to pull the heat flux value from the user object.

Refs idaholab/relap-7#484