Remove unnecessary Jacobian materials from ComputeMaterialsObjectThread
`ComputeMaterialsObjectThread` is only ever called during initial setup and
its only real purpose is to initialize stateful properties (despite the class
name which suggests it does something more). We use this thread to initialize
old values and it's sufficient to only do this through operation on the AD
residual materials.

Refs #5658