Removing `implicit_rdg` flag
The implicitness is now inferred from the time integrator, so users
do not have to set the flag in an input file any more.

The flag has to be set to false for ExplicitEuler, ExplicitMidpoint and
ExplicitTDV-RK2.  For ActuallyExplicitEuler, the flag does not matter
and it is the way how explicit will be done in future. So this is here
mostly for backwards compatibility and currently to be able to use RK-2,
which will be implemented in the new approach soon.

Closes idaholab/relap-7#554