Make phase specification consistent with PorousFlow2PhasePP
This makes this class consistent with the other formulations, including
making the calculation of capillary pressure consistent. The tests were
incorrectly using the new phase convention, so while the behaviour was
correct, the capillary pressure was the wrong sign and gas porepressure
was less than liquid porepressure. These tests have been regolded

Closes #12985