This illustrates that when a compute=false Material has stateful Material Properties, initQpStatefulProperties is not being called for Dirac Kernels that need the Properties.
ComputeFalseMaterial has vector and a scalar stateful Properties
UseComputeFalseMaterial just instructs ComputeFalseMaterial to compute those Properties

computefalse.i has a DiracKernel that uses the ComputeFalseMaterial properties.
Without the DiracKernel, the simulation runs fine:
 - initQpStatefulProperties is correctly called for each element in the mesh.
With the DiracKernel, the simulation crashes in the first timestep, as
 - the vector of properties is not correctly sized, because initQpStatefulProperties hasn't been called for the 'Dirac element'

Can someone fix this, please

I've tried to use the same kind of architecture as used in TensorMechanics.  I'm sorry if it's a little convoluted.

Refs #12545