Be more specific for Outputs page
Mastodon has a block Mastodon/Outputs, so when building its docs this page has problems locating which
page it should use. Ideally, this would be a fix that occurs in Mastodon, but I don't have a good way
to do that at the moment nor do I have any ideas to how it could even be done. So, for now just fix up
MOOSE and hope this doesn't continue to happen.

(refs #9638)