Don't try to do any ghosting activity with the AMG object until we can retrieve it
During initial mesh setup for DistributedMesh there is a call through to
ghosting functors. Our ghosting requires the existence of an AutomaticMortarGeneration
object, which is held by the FEProblemBase, which hasn't been created by the time of
this initial call. So if we don't have an executioner yet (which is actually setup
after the FEProleBase), then we just return and don't do anything