Enable running mortar on a displaced mesh
This commit was done to enable running the combined gap_heat_transfer_mortar_displaced
test with the new mortar implementation. As was true for the old mortar implementation
this only works for auxiliary displacement variables. E.g. we are not going to update
the mortar segment meshes during residual and jacobian evaluations yet, so we cannot
use non-linear displacement variables. Additionally the AugmentSparsityOnInterface
ghosting functor assumes that the element ids are equivalent between reference and
displaced meshes. This will certainly be violated at times when performing AMR, e.g
this probably won't work with AMR.