Change key for master_node_and_elem_to_xi1_slave_elem
Previously the key was a tuple of <const Node *, const Elem*>. This can
lead to different ordering of the map contents across different processes
which leads to a different sequence of element generation for the mortar
segment mesh because we iterate over this map when projecting master nodes
onto the slave face and splitting the mortar segement elements. The new
key is a tuple of <dof_id_type, const Node *, const Elem *> so that there
will be consistent ordering across processes.