Delete remote elements if its needed
Previously the CreateDisplacedProblemAction was telling the mesh not
to delete remote elements when adding the geometric ghosting functors.
During the addition of algebraic relationship managers we told the mesh
that its allowed to delete remote elements now...but then we never actually
call prepare_for_use again so the mesh never got distributed (e.g. we
never deleted the remotes). Now we've added an explicit task that will
manually delete the remote elements post EquationSystems init in the cases
where we told the distributed mesh NOT to delete remotes during the prepare_for_use
call

Refs #13145