Make sure that we clear the stringstream before loading
Since we reuse the stringstream (holding the backup), it'll already
have data inside of it. The new backup data will just be appended to
the end where we won't be able to access it because we'll have
an existing backup sitting in the front containing information sufficent
to reload the restartable data objects.

closes #13310