Fixing incorrect flow channel position
This slipped through testing, becuase when we test jacobians, we disable
integrity checking, so the code did not discover this misalignment. This caused
some entries to show up in an incorrect position in the Jacobian matrix. Since
the entries were small the Jacobian test still gave good metric comparing to
the FD Jacobian.