PorousFlow tests work in parallel
The changes involved tightening tolerances, setting abs_zero appropriately and getting rid of nonlinear+linear solver options that were crap in parallel.
Also, a few simulations - fracture_flow ones, for instance - do not work in parallel for the reasons stated in the tests file.
When using Actions and KT stabilization, i have to include a dummy UserObject to ensure the correct ghosting: Refs #13388

Refs #13255