Run mortar-mech test with PJFNK and not FD
Give a heuristic for approximation the dependence of the gap on the
displacement variables, which is the most important missing piece of the
Jacobian. Other things like the dependence of tests, normals, Jxw, phis,
etc. on the displacements are not included, so we still need to use PJFNK.
The Jacobian norm ratio is approximately 1e-2 to 1e-3, so I think that's
pretty good. With lu as the preconditioner, the linear iteration count never
exceeded 1-2.