Disable volumetric_locking_correction for face materials closes #13420
_Fhat was zero because of a multiplication by _coord in the process
of the volumetric locking correction calculation. _coord can sometimes
be zero for all qps in face elements.

I can't see any reason why we would want to use the volumetric locking
correction for face materials, so this can be addressed by simply
disabling it in this case.

Because this prevents _Fhat from being zero, there's no need for the
checks for it being zero, so I removed them.