"White-list" objects for automatic scaling value calculations
Previously we were "black-listing" objects for this calculation. E.g. we would
not compute NodalBCs if we were computing the Jacobian for automatic scaling purposes.
Now we have white-listed kernels, e.g. when we're doing our scaling jacobian calculation
we only call through to kernels as opposed to *not* calling through to other things.
This makes the code for maintainable because now if a new type of compute object is added,
then the adder doesn't have to remember to prevent the scaling jacobian calculation from
calling through to it