Update inline links to avoid interaction with inline command
It was reported that the following string wasn't parsing correctly in MooseDocs, it was tracked down to be the LinkInline class in core.py that was capturing up to the final ).

like in [fig:yieldsurface] (after [!citet](chiang1994anew)).

To fix it, the text within the [] for links is not allowed to contain additional [, which seems reasonable. Another way to fix it
is just to begin a new line with the '(after...'

(refs #13611)