Don't let automatic scaling destroy the sparsity pattern
We call through to only kernels during the automatic scaling
calculation. These kernels do not capture the sparsity pattern
needed for things like DGKernels and InterfaceKernels, so if
we don't tell PETSc to not change the sparsity pattern, it will
shrink the allocation (I believe). Then when we actually call
through to compute the real Jacobian, we have to do a bunch of
mallocs in MatSetValues which makes the simulation super slow.
To fix this, we use @fdkong MatResetPreallocation function.

This also deletes FEProblemBase::computeResidualType which was
not being used anywhere.

Refs #12601