Refactor where scaling data lives and is computed from
This commit moves calculation of the scaling Jacobian from the
beginning of NonlinearSystem::solve to PicardSolve::solve. This
ensures that the Picard norm will match the initial non-linear
residual norm of the master application. Additionally all scaling
data members were moved from FEProblemBase and NonlinearSystem
into NonlinearSystemBase. It just seems appropriate that the
nonlinear system (where traditional nonlinear or eigen in the future)
owns the data related to scaling of the residual and Jacobian.
FEProblemBase to me is an object that should be designed to hold
the data related to using finite elements to create an algebraic
system. The nonlinear system then solves that system using Newton
technology (typically)