Do some RelationshipManager duplication
We want to have smart sparsity specification through our ghosting
functors. This means using our custom CouplingMatrices which
know the user-specified coupling between variables. This is
relevant for nonlinear-systems so we want to make sure we are
attaching that coupling information to ghosting functors operating
on the DofMap associated with the non-linear system. Obviously,
we cannot use that same coupling info for the aux system. There
it doesn't make any sense. Hence we must duplicate
RelationshipManagers; when the user requests an RM that is
evaluable and coupling, we must create two RMs: one that will
be attached to the non-linear DofMap that contains non-null
coupling matrices, and another that will be attached to the aux
DofMap that doesn't act on coupling information