Really remove libMesh default ghosting/coupling
We are now fully explicit in the ghosting and sparsity coupling
pattern that we want. We add a coupling functor for kernels and
bcs that does not do any ghosting because it has _n_levels set
to 0, but it does specify intra-element dof coupling so that the
sparsity pattern is correct.

Now we can really re-gold the geometric edge test case because
there is really no evaluable ghosting now