Address failing recover test
How to deal with this test? If the user has requested automatic
scaling once, then we really should not be scaling upon recovery.
We should use the factors from the previous portion of the sim.
In order to do that, the scaling factors need to be restartable
data. however, in order for the data to be restartable,
MooseVariableBase needs to inherit from MooseObject. So maybe
now is finally the time to make MooseVariableBase inherit from
MooseObject???