Remove code using gmsh api
Gmsh code is licensed under the GPL license. In theory we
could make a directory of scripts like this licensed under
GPL and not worry about it since these scripts cannot be
linked to or included for library purposes. But that's some
work, so we might as well just remove.

Refs #000