Make sure that 1 layer coupling is actually added
This marks the importance of local testing. 1 layer
of sparsity was not actually getting added in a Rattlesnake
simulation that added all its dg/interface kernels through
actions. The issue was that the logic in Action::addRelationshipManager
had determined that because the added RM had an RM type of
GEOMETRIC | ALGEBRAIC | COUPLING that the RM must have been
added during the geometric attachment stages and hence did not
add it for the requested `input_rm_type = COUPLING`. Now in
our `CouplingFunctorCheckAction` I make sure that we add an
RM with rm type restricted only to coupling, as the name of
the action implies. Note that this logic presumes that the user
has added a relationship manager that provides geometric
ghosting ahead of this coupling ghosting. If not, the user will
be in for a rude awakening if they try to run with DistributedMesh.
Note that we can't try to fix this situation for them; by the time
we've added all our compute objects to the warehouse, the mesh
has been prepared and remote elements have already been deleted.

Refs #13736