Show geometric ghosting failure
This requires first actually turning off default mesh
ghosting in MooseMesh. I will need to reconfigure this
to somehow be user settable as default_ghosting is a param
in FEProblemBase. The issue here is that the FEProblemBase
object is created after the mesh has been prepared and remote
elements have been deleted.

Anyways, when we remove default mesh ghosting, we show that
when some object requires algebraic ghosting the simulation will
fail because algebraic ghosting requires corresponding geometric
ghosting. And unfortunately the CouplingFunctorCheckAction just
runs too late to provide geometric ghosting. This means
the responsibility falls on the user