Move MaxVarNDofsPerElem calculation
Previously this was in NonlinearSystemBase::init
which gets called before our computing objects get added.
The problem with this location is that it didn't allow
us to run the MaxVarNDofsPerElem calculation intelligently.
That object exists only for AD purposes, hence if we can
move its calculation to NonlinearSystemBase::intialSetup
then we can see whether we actually have any AD objects. If
we don't then we don't need to run the MaxVarNDofsPerElem
calculation. This was showing up as the primary computational
expense in one rattlesnake simulation, where it served no purpose.

Tagging @YaqiWang because he should appreciate this speed-up