Initialize _fe_type and _component in constructor
Marmot has an action which inherits from AddAuxVariableAction
and which makes use of the _fe_type and _scalar_var members
in its `act` override. For backwards compatibility, we can't
rely on the `act` override calling `init` so we have to make sure
that _fe_type and _scalar_var and init'd in the constructor.
Again we are making use of the fact that we've duplicated params
between the MooseVariable* MooseObject and the AddVariableAction
action object