chigger: allow to overwrite previously rendered movie
Adding option to allow to overwrite a previously rendered movie.
When scripting movie rendering, we may not want to wait for user
to answer the question if the existing movie file should be
overwritten - ffmpeg does that.

This is simply accomplished by passing '-y' flag to ffmpeg which
will cause that all questions asked by ffmepeg will be answered by
'yes'.

Closes #13924