Use the new MOOSE API to add variables
Changing the structures in Simulation to be able to use the new MOOSE
API to add variables. Now we do not use subdomain IDs anymore, but we
use subdomain names. The component facing API had to change as well.
In components, we store both IDs and domain names, so we now use just
names and drop IDs. IDs are only used for mesh generation now.

Refs idaholab/moose#10601