Ensure that _ad_JxW is always properly sized
_ad_JxW needs to be resized whether we are reiniting
on the face or in the volume. In the former case, it's
computation is meaningless but it's still done, so we can't
have it be the wrong size.

Also add a lot of comments about what the AD mapping methods
are doing in the Assembly class. This should be helpful to
developers (myself included) trying to figure out what is going
on