Restore stream settings after TimedPrint/PerfGraph
This bit us in the test added from #13798. If we generate output
from TimedPrint (e.g. if mesh building or equation systems initing
takes **enough time**), then we set the precision to 2, and then
this impacts unwary users who attempt to use the console and
expect default precision (e.g. 6).