Do not evaluate nodal BCs for non-nodal variables
In SystemBase::reinitNode{Face}(), we checked if the variable was nodal
and if it was we called reinitNode() and then computed values.  This caused
that the internal state of MooseVariable was not updated, so when we got
into a nodal system and asked the variable if it was defined at nodes via
isNodalDefined() call, the answer was incorrect.

The SystemBase::reinitNode{Face}() needs to update the internal state of
MooseVariable first and then if the variable is defined at the node, we
compute its values and possibly evaluate the objects associated with the
variable.

Closes #14019