More careful MeshGenerator checks
When checking for whether we should set the MooseMesh type to
MeshGeneratorMesh, we need to be more careful about checking
which `Actions` have triggered the `add_mesh_generator` task.
If it is a custom action, they may not always be adding
mesh generators depending on user input, so we need to have
more careful checks when both setting up the mesh and when
setting the `MeshBase`.

Refs #14049