Slacken neighbor check in ComputeMaterialsObjectThread::onInterface
This is actually necessary for any calculation with
neighbor stateful material properties. Interface objects
are identified with one unique side unlike InternalSide objects
which could be identified with either face. In the interface
case the elem_id < neighbor_id comparison is actually very bad.
If the element that has the face corresponding to the user specified
sideset has an ID greater than that of the neighboring element, then
the neighbor Material will never initialize its stateful properties