Get variable corresponding to correct thread ID
The PorousFlowFluidMass postprocessor was getting its _var
member from the PorousFlowDictator GeneralUserObject. Through the
logic in UserObjectInterface::getUserObject, this returns the
thread 0 copy of the PorousFlowDictator. This thread 0 copy also
has thread 0 copies of the variables through the Coupleable
interface. If the thread 0 has not
reinitialized its variables before the thread 1 copy of
PorousFlowFluidMass runs its execute routine, then the value
of _var->phi() will be garbage. Hence we must ensure that we
are getting a pointer to the correct thread copy of _var

Refs #14151