Allow convergence on zeroth nl iteration for ReferenceResidualProblem
I don't understand why the original author made it so that
convergence could not be achieved on the zeroth nl iteration.
This removes once of the principal advantages that I see of
ReferenceResidualProblem in that it can recognize essentially
that we've achieved a steady-state condition and/or that it
doesn't make any sense to keep iterating because the variables
are converged with respect to reference quantities.

Refs #14004