Use available xi1_a and xi1_b to determine mortar segment
Previously when trying to determine what mortar segment a
master node projected into, we did a distance check that
is predicated on using linear elements. However, it is easier
and more robust to just see whether the xi1 from the projected
master node lies between the xi1_a and xi1_b values of the
mortar segment.