Reduce number of DualNumber operations for undisplaced problems
Use regular versions of _JxW, _coord, and _grad_test where
appropriate. One sad thing is that if a user inherits from
ordinary ADKernel and uses _grad_test then they will be
performing a bunch of unnecessary DualNumber operations
if they are computing on the undisplaced, but I think it's
better to be correct than to be fast. I tried to think of
clean solutions for making it so that _grad_test is always
the right thing, but that would require template arguments
whether in the class or in the method in order to make it work

Closes #13027