Update rom_stress_update heavy verification test
The creep_rate_diff Postprocessor seems to be a
very sensitive quantity. I'm kind of a firm believer
in not doing diffs on anything except non-linear
variables and/or their aggregate quantities because
those are the dofs that the non-linear solver is trying
to crank on. The change in method for calculating scaling
factors seems to have very slightly changed the scaling
factors. However, if I do exodiffs on the non-linear variables
they come out clean, so I think I trust the change. So I've
switched this test to be an Exodiff and the only quantity it
doesn't test is creep_rate_diff. So creep_rate itself is fine.

Refs #14204