Reduced active nonlinear function set (RANFS) scheme
A few lessons learned while implementing this algorithm:
Like Barry says, you should never alter your function
just because you're computing a linear vs a non-linear residual.
If you are doing that, then you are doing something wrong. So I
learned that I should not just do the computation of the lagrange
multiplier during the nonlinear residual evaluation because then
when doing JFNK there will appear to be no dependence of the master
residuals on the lagrange multiplier calculation which is function
of the slave node residual which is a function of all the dofs
on all the connected elements

The added test should always converge in two non-linear iterations
because aftr the first iteration we identify the active set,
and then we solve with the reduced residual set during the second
non-linear iteration and then we are golden

Refs #14454