Actually base phi_size off the dof indices
If we don't we run into this issue: we call copyShapes for
the jvar_number which just ends up copying in the shape functions
corresponding to the basis functions for that variable. Those
shape functions are calculated in the Assembly class at the
beginning of element residual computation, e.g. FEProblemBase::prepare
-> Assembly::reinit, entirely divorced from variable degrees of
freedom. Hence the phi obtained after copyShapes absolutely will
have a size even if the variable doesn't live on the undisplaced/displaced
mesh