Set execute_on for TagVector(Matrix)Aux to timestep end
These aux computations rely on the nonlinear variables already
being evaluated so we have a dependency issue if we allow these
things to be evaluated on linear or potentially other execute
stages. AFAIK the only real use case for these computing objects
is for "visualizing" reference quantities akin to save-in values.

Closes #14427