Prevent numerical garbagge to kill the run
Numerical garbage can produce alpha_liquid value that is outside of the
allowed range. This that causes the run to terminate.

Adding a cutoff region that is smaller than the width of the exponential
region in the VolumeFractionMapper, such that if the remapped value of
alpha_liquid is outside of the region, it will get moved inside the region.

Also, removing alpha_vapor_bounds, because this information is already stored
in the VolumeFractionMapper as _minimum and _maximum.

Closes relap-7#731