Assume units
When users select a unit group, the output unit select by default should
be the one used by THM/RELAP-7, so users do not have to hunt for it.

Also, reordering the units so that English units are first and will be
therefore selected as an input unit.  The reason is that this plugin
will be used in first place for converting from English units to SI unit.

Refs #19790