Fixup usage of enumerate utility
Currently, if somebody compiles MOOSE with -std=c++17, they will
get errors from use of Moose::enumerate. This is because the utility changes
the return type of enumerate conditionally based on the compiler standard
in use, which is a really poor idea. This PR returns the C++11 compatible
shim class, which won't quite work with c++17. We'll have to change the
calling code anyway in the future.

refs #13126