Update the Conda instructions
Include more detail on moose-environment removal.

Use multiple commands to add channels. I and Logan seem to run
into an issue when instructing conda to add multiple channels. We
think it has to do with zsh?

Remove the line about building moose after building libmesh, as
the nect section is building moose, and is implied that it is
the next step.

Refs #14515