Add no_error_deprecated to the SM variants of combined tests ref #6978
These tests all have tensor_mechanics equivalents, and are being
retained until solid_mechanics is removed to ensure that the
SM code is still tested until it is removed.

Also converted the glued_contact_constraint test to TM. It was the only
test that did not have an SM equivalent.