Fix --failed-tests, Unittests
Create unittest for --failed-tests. While doing so,
discovered a rather nasty bug:

Fix the potential for unittests to fail, and not reporting
it. We were returning error codes, but not setting the
status of the test

Allow TestHarness to use a different naming convention for
.previous_test_results.json. This was necessary to prevent
the many-failing-tests in unittests from clobbering the
results file during the crucial --failed-tests test.

Jeez... fix all the dumb naming conventions of results_file,
results_storage... was overly confusing with similarly named
variables. results_file is a file object. results_storage is
json object.

Closes #14512 #14575