Fix PATH issue in translator test
Not knowing much how MooseDocs testing works, it would seem the
current path is not correct when the test begins (CWD is in the
test directory, not MOOSE_DIR/python).

The proper method I think, would be to use the ABS path to the
config.yaml file. However, I am not sure how to obtain that within
the unittest (MOOSE_DIR/python/MooseDocs/test/config.yaml).

Closes #14588