Fixing design problem
Removing parameters from the c-tor of VolumeJunctionBaseUserObject to allow
the implementation of VolumeJunction with NCGs.  The problem is that we need
to know how many NCGs we have, which comes via the coupling API, but the
method that should give back the list of variables is called before the object
is constructed, thus we are unable to use `coupledComponents` call.

Refs #19784