Create 'parameters' package to be use by peacock, MooseDocs, chigger, TestHarnes, ...
This package is an improved version of the "Option/Options" objects in chigger that were developed
to mimic the current TestHarness parameters but be more robust. The two versions exist because
chigger was at one time a stand-alone application. This is the beginning of going to a single
tool for all the parameter/options across the various projects

(refs #12095)