Use _fe_problem to reinit higher dimensional elements
I choose to do this for a couple of reasons:

1. If using _fe_problem then I don't have to change the API
of SwapBackSentinel.
2. This is much safer. This allows the MortarConstraints to use
material properties  coming from both displaced and undisplaced
materials. It's hard for me to imagine many scenarios where a user
would want to use an undisplaced material with a displaced constraint
or visa versa, but one example would be if using a
GenericConstantMaterial whose property  values are totally unrelated to
whether the mesh is displaced. It would be odd to require the user to
match the displaced setting of the constraints in this case. There is a
drawback here though, and that is speed. With this change we're going to
calculate variable values on both undisplaced and displaced meshes
regardless of whether they're needed. There's probably some checks we
can do in future commits that can limit computations done to
computations that  are needed. One idea I had was to do undisplaced and
displaced mortar constraints at the same time because then we could
potentially cut down on reinits by a factor of two...but this doesn't
really fit well into the current framework because we're looping over
mortar mesh elements (which will have fundamentally different
structures depending on whether the parent mesh is displaced or not)