Allow interface kernel execution on gmsh sidesets
Do a simple check in the `InterfaceKernel` for whether `_var`
is active on `_current_elem->subdomain_id()` and for whether
`_neighbor_var` is active on `_neighbor_elem->subdomain_id()`.
This prevents the IK from being executed for the wrong sidedness.
This check should eventually be replaced by something that happens
ahead of all the reiniting in things like
`ComputeResidualThread::onInterface` but I don't have the time to do
that at present.

Closes #15023