Shift dependencies
There is some sort of new conflict with Mac, and the way I am setting
dependencies for moose-compilers/moose-mpich. Instead of having moose-mpich
depend on moose-compilers, have both contain the exact same set of dependencies
by created a symlink between the two conda_build.config.yaml files.

Why this works, but not when they depend on each other, is a mystery to me at
the moment.

Shift the libpng dependency into libMesh. As it makes a bit more sense having
that there, rather than in compilers/mpich. It may even belong in moose-tools,
but moose-tools is not required to perform run_tests, though, libpng is.

Refs #15058