Change MaterialProperty value storage type
It used to be a MooseArray, but I changed it to std::vector in
402ef243c40c because I needed to be able to pass construction
arguments to the stored typed being created (I can easily do this
using std::vector::emplace_back). *But*, now that MooseADWrapper
is just a typename and not a class which takes an argument, I no
longer need to be able to pass construction arguments, and I can
change this back to a MooseArray. This enables conveniences like
storing either a variable value or a material property in a given
symbol, e.g.:

```c++
MooseArray<Real> * foo;
if (use_var)
  foo = &_u;
else if (use_mat)
  foo = &mat_prop.get();
```