Deprecate BOOST distributions ref #12688
Point the deprecation messages for all the Boost distributions to
the appropriate non-Boost distribution.

I had to move the implmentations for some of the Distribution classes
into a separate class, and add a wrapper Distribution class because of
difficulties in using both registerMooseObjectReplaced() and
registerMooseObjectAliased() together. This gets a head start on
moving all of these over to classes without Distribution at the end
of the name anyway.