Don't bypass Material::get(AD)MaterialProperty
Before this commit `getGenericMaterialProperty` was skipping
straight through the `Material` class and going straight to the
`get(AD)MaterialProperty` methods in the `MaterialPropertyInterface`
class. This meant that for one thing `_requested_props` wasn't getting
set in the `Material` class, which meant that dependency resolution
was incorrect for any `Material` that employed
`getGenericlMaterialProperty`. With this change, dependency resolution
is correctly restored