fix face info tests for 64 bit dof ids.
We were using maxint as a special number for diffing against in these
tests - this obviously doesn't work when the special number is not
static across all testing configurations.  So I just changed to use -1
instead.  Regolded and now it works.  This fixes the 64 bit failures
introduced by #15049.  This was happening when either the elem or
neighbor element weren't defined on faces where a variable was only
defined on one side.