Small number of cores for partition
The number of cores for the graph partition can be different from that used
for mesh generation and simulation. Partitioners often become inefficient in
compute time when the number of cores is large (around 10,0000).
A possible "fix" is to partition the graph using a small number of cores
when the mesh generation and the numerical simulation use a large number of processor cores.

Closes #15464