Revert "Merge pull request #15118 from austinltam/fragment_cut" ref #15115
This reverts commit 757b6dbff9f04bd754adc78c12abf5cfaa3a10de, reversing
changes made to 01ad7fec9f41f8f2e4329ebbfc62648d4a474d73.

This code introduced a memory leak and is holding up merges to devel.
I'm in the process of converting the code to use shared_ptr, which will
resolve this issue, and will add this code back in at that time.