Remove Conda update instructions
Remove the bit about keeping Conda up-to-date. This is being done (and needs
to be done) during the main 'Getting Started' page, when we go to mention
_both_ Conda and MOOSE repository. They need to be done simultaneously, and
it doesn't make sense to teach users how to do that during a Conda install,
before they even have a moose repository to update.