Ensure block-restricted save-in consistent for different proc counts
This was a pretty interesting bug. One processor owned all the elements
for a subdomain that the save-in variable was restricted to. But we had
a node on a processor boundary that wasn't local to any of the subdomain
elements. Consequently when executing the `AllLocalDofIndicesThread`
with local elements, no process will try to zero that variable's degree
of freedom corresponding to the node on the processor boundary. If we
instead just add all dof indices that are semilocal to the local element
(e.g. all the dof indices attached to the local element) and zero, then
we ensure that we cover this corner case.

Closes #12239