only load system vectors that existed in the old simulation
When "restarting" a transient simulation from a steady-state simulation,
the transient systems will have more vectors than the steady-state
systems. Naively trying to read data corresponding to all the vectors in
the transient simulation will generate garbage results because you're
trying to read-in data that doesn't exist.

Fixes the first part of #15333