Adding a getter for fluid properties
Add a getter to access fluid properties used in a flow channel associated with
the heat transfer component.  This is useful when a closure class wants to setup
a material for heat transfer coefficient and needs to know the used fluid properties.

This information is already stored in the class, so all that is needed is really
a getter.

Refs #19682