Don't remove default ghosting from the mesh for DisplacedProblem
For quite a while now we haven't removed default ghosting from the mesh
from `FEProblemBase`. If we have `ProxyRelationshipManagers` for
instance, we rely on having identical ghosting from the reference mesh
to the displaced mesh (indeed we should have this no matter what). If
default ghosting is missing from the `DisplacedProblem` but not from the
`FEProblemBase` then there is a
very strong chance that ghosting will not be identical between the two.
This is certainly true now that I've added periodic boundary ghosting to
the default libMesh `GhostPointNeighbors` ghosting functor.